adjust network test and move to new network part 2 test
